backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
email-validator
passlib[bcrypt]
sentence-transformers
pgvector
python-jose[cryptography]
python-multipart